ankiclient: adds cleaning of temporary files
The {audio-media}, {screenshot}, and {screenshot-video} markers created
temporary files in the temp file directory and didn't delete them
because I assumed any OS would do that for me in a reasonable amount of
time. Enter Windows, an OS that keeps temp files until the heat death
of the universe. Now temp files are deleted once Anki returns a reply.