mainwindow: fixes crash when loading a file with --secondary-sid
When loading a file with an secondary-sid=/=no, graphical jank would
occur. When changing the secondary subtitle track in the UI, a crash
would then occur. This commit returns to the old way of doing things
where each QAction handles things in its own lambda. This simplifies
the code greatly.