playeradapter: adds tempaudioclip method and removes ffmpeg dependency
Previously the way the {audio-media} marker was implemented was going
outside of mpv to a modified piece of ffmpeg example code, transcoding
what was hopefully the right audio track, and adding that to the Anki
note. This worked for the most part but after adding official support
for streaming, this was not sufficient.

The new way of going about it is using mpv's built in encoding
abilities. This is how it always should have been since I could use a
high level API rather than the low level libav API I was using before.
This makes everything much simpler and less error-prone.