mainwindow: reverts 1000px minimum width for player, makes dynamic
Reverts the 1000px minimum width for MpvWidget back to 500px because it
caused problems when SubtitleListWidget was open. It also prevented
cleanly resizing the player to take up exactly half the screen by
dragging the window to the right/left side of the screen. This makes it
so the player has a minimum width equal to that of the controls minimum
width in fullscreen.