dictionary: fixes bug where mecab error message would not appear
In cases where mecab couldn't find its system dictionary, a MessageBox
should have appeared warning that mecab was not functioning. This didn't
work because it was an error triggered in the constructor and relied on
MainWindow handling the signal for showing a critical error message. The
prefered method for showing MessageBox errors from constructors will not
be invoking the static QMessageBox methods.