dict: major optimizations to searching
I don't know what I was thinking when I orginally wrote the concurrency
code for the dictionary database, but it is the most boneheaded thing in
the world. Previously each search would cause needless reads to disk to
get the dicitonaries and their priorities. Now all of that is cached in
memory and only modified when the dictionaries change, which should be
almost never for most people.