player: moves osd on top of player
Completely refactors the OSD code, taking most of it out of MainWindow
and segmenting each piece into its own file. The menubar duties were
taken from MainWindow and put into a custom widget called PlayerMenubar.
This has the benefit of having a working autoFillBackground() method,
but comes at the cost of having the menubar the same color as the title
bar on KDE. Most people will not be affected by this change though, so
its worth it.

The player OST now also has fade in and out animations to make the UI
seem more seemless.