cmakelists: adds x86_64 cross compile target for mac, bundles ipadic
Adds a cross compile target for amd64 macs to compile x86_64 binaries.
It seems like a universal binary is out of reach due to Homebrew not
offering universal binaries, so versions build for different archs will
have to do. Also bundle in ipadic so app bundles are portable on macOS.