iconfactory: fixes memory leak
The destructor in IconFactory set m_factory to nullptr on being called.
In buildIcons, a StyleFactory is allocated on the stack on Linux,
leading to the destructor being called on returning from buildIcons.
This set m_factory to nullptr, making the next call to
IconFactory::create() allocate a new IconFactory.

Getting around this, IconFactory::destroy() deliberately sets m_factory
to nullptr after calling delete on.