cmakelists: changes macos app bundle build system to fix local deps
Building app bundles on macOS now requires that dylibbundler be
installed because macdeployqt missing a ton of depedencies. dylibbundler
does a much better job, but it still misses Python, which is depended on
by libvapoursynth-script. Through some tricks, I was able to make it so
that Python can be added regardless of what the version is or what
arch is being built on macOS. The CMakeLists are not very readable, but
they are heavily commented.