subtitlewidget: fixes high cpu usage
The previous implementation of SubtitleWidget had very high CPU usage
for unknown reasons. Turns out that an infinite loop was created by
overriding paintEvent and then doing cursor operations in there.
Basically new updates were being created inside of paintEvent which
would baloon over until subtitleWidget was hidden or the event queue was
too large.

This fixes the problem by completely ignoring the StackOverflow solution
and instead doing things the "dumb" way, i.e. having one QTextEdit with
a stroke in the background, and another QTextEdit overlayed on top of
that QTextEdit as the foreground. This achieves a similar effect without
the high CPU usage or having to override paintEvent.

This new widget is called at StrokeLabel and both ProgessSlider and
SubtitleWidget depend on it.