strokelabel: prevents the widget from being resized larger than parent
fitToContents previously would resize the StrokeLabel to whatever size
it wanted to be, but this would have unintended effects such as forcing
the window to be bigger if the stroke label needed to be larger than its
parent. This bound the dimensions by the size of the parent.

Design was also improved to prevent fitToContents from being overridden
and making the internal QTextEdits private. This has an effect on
SubtitleWidget which used to override fitToContents.