strokelabel: revert 4bb801f
Causes stroke labels to get stuck in updatesEnabled(false) mode if text
is set in the transition from fullscreen to windowed and vice versa on
macOS. Turns out that the implementation of updatesEnabled() in Qt sucks
because once it is set in a child widget, it cannot be unset, i.e. go
back to mirroring the value of the parent automatically.

The piece of code is nonessential, and was only there to improve
performance in some cases, so reverting is no big loss.

Closes #44