subtitlelistwidget: fixes crash when re-adding subtitles
When subtitles would be restored from the subtitle cache, it was
possible was a crash to occur due to subtitles getting double added to
the subInfos list. This means it is possible the for-each loop to return
a NULL reference, leading to a crash.

This fixes the problem by making sure restored subtitles are not
re-added to the list they are cached in by adding a new helper method,
addTableItem. Both addSubtitle and handlePrimary/SecondaryTrackChange
depend on this method to add subtitles to the list.

There is also some added synchornization code just in case.