subtitlelistwidget: prevents double adding of preloaded subtitles
The start time mpv determines a subtitle has is possibly different by
some margin of error than the value Memento determines it has after
parsing subtitles. This just adds an error value to the start time to
determine if Memento has seen the subtitle before. This should prevent
most if not all instances of double adding of subtitles.