subtitlelistwidget: rework parsed subtitle display
This commit eliminates additional reliance on mpv sub-text properties
when updating the subtitlelistwidget, and offers a settings option to
turn off flattening (compressing) simultaneous subtitle entries.
subtitlelistwidget now maintains a flag to show whether it has parsed
subtitles & populated the list. A new getTimePos property is added to
mpvadapter. Now on subtitle change, if the flag has been set,
subtitlelistwidget assumes that the newly appeared subtitle is already
in the list, and looks it up with timePos. addPrimarySubtitle have been
refractored to updatePrimarySubtitle for this purpose (& secondary).