subtitlelistwidget: refactor and bug squash
Refactors SubtitleListWidget to encapsulate all the items for primary
and secondary subtitles in their own structs. This makes maintainability
massively easier because structs can be passed around as opposed to
verbose C++ types.

Fixes bug where secondary subtitles would have the subtitle regex
applied to them, despite that not being intended behavior. Now the
subtitle regex is only applied to primary subtitles.

Fixes any forseeable bugs with parsed subtitles not being selected if
multiple are on screen at once. Takes advantage of the fact mpv emits
subtitles that have overlapping timecodes seperated by newlines. Does so
by indexing into a MultiHash using each line to find a corresponding
QTableWidgetItem. This item can then be used to find the corresponding
SubtitleInfo. From here all that needs to be checked is that the current
time is between the start and end times of a SubtitleInfo. This is very
efficient since it only depends on the number of newlines in an mpv
subtitle. This did require having a non-regex filtered version of the
subtitle emitted from MpvWidget because a regex filtered subtitle may be
missing newlines inbetween compressed subs.