subtitlelistwidget: fixes potential dangling pointers
Originally the implementation would keep the address of the result of a
QHash. Turns out this isn't safe since if a QHash get's resized, that
address may no longer be valid. This dynamically allocates the values
before putting them into the QHash so everything is guarnateed to be
safe.

That said, the current implementation lets QList handle the allocation
and deallocation of SubtitleInfo structs. This is safe in Qt5, but when
Qt6 comes, this will no longer be safe because QList will be replaced by
the implementation of QVector. Although not urgent, it is something to
think about.