subtitlelistwidget: fixes regex subtitle caching corner case
In the event embedded primary subtitles are use, the subList is
populated with regex filtered text. If the track then becomes the
secondary track, the cached subtitles will be regex filtered, which is
incorrect. This will lead to duplicate entries being intereted into the
subList. This prevents this issue by only regex filtering the
QTableWidgetItem from embedded primary subtitles.