fonts: changes noto sans cjk jp to noto sans jp
Google changed the name of their CJK fonts at some point to drop the CJK
from the name. This is a problem because it created dead links and
caused computers to fall back on the wrong fonts. This is mostly a
problem for Windows users since their font rendering sucks.