subtitlelistwidget: fixes strange shift selection behavior
The new way of selecting subtitles for embedded subtitles had a bug
where when selecting mutliple subtitles with shift, the incorrect
subtitle would be selected in some cases. Turns out that shift-cicking
starts selecting after the current item, not the earliest selected item.
This changes makes sure the earliest selected item it always the current
item.