definition: large refactor of definition widget
The larger point of this refactor was to allow definition widget to be
repopulated with terms without first being destroyed. This would allow
easily for the creation of a yomichan style search function since terms
could be updated constantly.

The refactor makes use of shared_ptr in order to make multithreading
AnkiClient easier and prevent issues with use-after-free in those
difficult situations.