searchwidget: adds the ability to search for terms
Allows the user to open a search widget in order to search for terms
that don't appear in subtitles and then make cards as if they were in
subtitles. The caveat to this is that cloze may not work as some expect
since it is derrived form the query, not the current subtitle.

This change required a ton of refactoring since the DefinitionWidget is
no longer unique. Essentially mangement of visibility signals was
offloaded onto PlayerOverlay so that the two DefinitionWidgets could
work independently of one another.

Another change is sentence, start, end, and titles are added at the
point of adding a card. This will work well in most cases, but if
sentence is used on the front of cards, it may behave unexpectedly. This
can be solved with a change of the duplicate policy, but that may be
unituitive. I don't know if anyone actually uses sentences on the front
of their cards, but I don't see a good way to handle that case since
cards can be added independant of whether or not the current subtitle
changes. Using {context} would be best for these use cases.

Since the SubtitleList and SearchWidget are seperated by a splitter,
some hacks had to be implemented to make sure the splitter that
seperates the player and subtitle list is invisible when neither the
SubtitleList and SearchWidget are invisible. Those hacks were
implemented in MainWindow.

To toggle the visibility of the SearchWidget, you can go to Settings >
Show Search or press Ctrl+R.