dictionary: single threads appending to list in searchterms
The previous implementation multithreaded appending to a shared list
that would eventually be returned. Although this doesn't seem to be
a problem, there is a rare case where on macOS QList will attempt
a realloc and fail due to a misaligned pointer. I don't know if this is
a bug on QList or Clang's standard library, but single threading
appening to the QList seems like the safest option.