ankisettings: add option to normalize media audio
Adds option to normalize the audio of {audio-context} and {audio-media}
to some user defined dB value. Uses ffmpeg's loudnorm filter with a
default value of -20.0 dB. This is a little loud for my tastes, but it
chosen due to its similarity with JapanesePod101's audio levels.
Comparing nomalized audio to {audio} does demonstrate similar audio
levels.