kanjiwidget: fixes macos where desktop would show on searching kanji
If the user manually searched a kanji by clicking it on a TermWidget,
was using macOS, had packaged Memento as an AppBundle, and was in
fullscreen mode, the desktop would be shown upon searching. I have no
idea why this happened, but it appears to have been triggered by calling
setVisible on buttonBack.

To be honest, this is the most bizarre bug with the strangest conditions
to cause to happen that I can't help but wonder why it happens. It makes
no sense whatsoever.