mpvwidget: sets screenshot directory to desktop by default
Previous behavior had screenshots saved wherever Memento was launched.
This is fine on Linux because that is usually the same directory as the
file, but it becomes a problem on Windows where Memento is always
launched in a write protected directory.

The new behavior is in line with mpv's default pseudo-gui behavior of
saving screenshots to the desktop. Reminder to walk a finer line between
accessability and extensibility in the future.