subtitlelistwidget: removes reexecuting a search on tab change
This isn't the best idea because searching changes the selection. They
may have wanted one selection to be the result of a search, but for the
other set of subtitles, they may want the selection to be the current
subtitle. Removing the automation from the feature makes the choice
explicit.