ankiclient: modifies {pitch-graph} to use text color instead of b&w
The previous version used black and white pitch graphs which had the
advantage of looking nicer if text was black, but if the background of
cards was black, they were useless. This changes pitch graphs to use the
text color for the stroke and tranparency for fill.

This required making the connecting lines shorter so that they aren't
visible under transparency.