subtitlelistwidget: adds shortcut ctrl+shift+c to copy audio context
Allows the currently selected audio context to be copied with
ctrl+shift+c. While this does 'work' for the secondary subtitles, the
current audio track is always used. What is copied to clipboard is the
url for a temporary file. The temporary file is deleted when something
new is copied to clipboard or Memento is closed.