playeradapter: allows secondary subtitles to be visible at all times
mpv added the ability to toggle the subtitle visibility of secondary
subtitles independant of primary subtitles. This finally allows
secondary subtitles to be visible at all times as a normal user would
expect without workarounds.

This feature is #if'd out of versions of mpv whose API is before version
2. This isn't going to be a perfect heuristic, but if you're running
master versions of mpv, you get what you deserve. It should work fine
when mpv eventually releases 0.35.0.

Due to sub-visibility merging secondary-sub-visibility into it, the
logic for only changing the visbility of the primary track becomes a bit
more difficult. This means there are two cases, secondary subtitles are
either visibile (yes, secondary-only), or they're not (no, primary-only).

PlayerAdapter accounts for these cases and changes the
get/setSubVisibility to only query the primary subtitle visibility as is
expected in Memento.