cardbuilder: shows suggestions when typing anki markers
This should help some people out when they're trying to figure out what
markers are available in Memento. This was implemented by sub-classing
QPlainTextEdit and using that for rows entry rows instead of
QTableWidgetItems. As a byproduct, newlines are now allowed, though
those won't render in Anki because input is HTML.