audiosourcesettings: adds support for yomichan-style json sources
Adds support for Yomichan style JSON audio sources. This means there are
now two types of audio sources, "file" type audio sources and "json"
type audio sources. File type sources simply mean they link directly to
a file. JSON sources on the other hand do not link directly to a file,
but instead to a JSON endpoint that describes what files are available.

For this change, a simple combobox was added as a column in
AudioSourceSettings.

To add support for this on the backend, a lot needed to be changed
inside of TermWidget. The changes made mean that AudioSources are now
cached inside of TermWidget, instead of sharing the array with
DefinitionWidget. This should protect against any undefined behavior and
make everything a little more resilient at the cost of some speed. When
playing audio or adding a card, if it is detected that there are JSON
audio sources, all JSON audio sources will be loaded in, and then normal
work will resume.

Also added is the ability to select the audio source to add to Anki
cards by right clicking the Add Note button. This works the same way the
Play Audio button works, except it adds a note. Because of this feature,
selecting the default audio source was removed. I always suspected this
would become a buggy feature if user audio sources became desynced with
Anki audio sources. Since there are ways to work around it's removal, I
see no reason to bend over backwards to try and make it work.