termwidget: fixes race condition that added incorrect card info
In the case where the user is using JSON audio sources, it is possible
for the TermWidget to be deleted before all JSON audio sources are
returned. This would prevent the current term from being added if they
were to quickly execute another search, or if they played media, the
wrong information could be added to cards.

This is an unpleasant way to get around that that has TermWidgets
schedule themselves for deletion once they no longer have an outstanding
network requests pending.