glossarywidget: adds support for structured content in terms
Adds support for Yomichan dictionaries that use StructuredContent in
terms. This is a massive feature that is unforunately hampered by the
limitations of Qt's ability to render HTML and CSS. Regardless, support
for inline images, tables, bold text, etc. is all possible on the front
end. The downside is <ruby> tags are still unsupported and using ems to
scale images isn't possible. In fact, scaling images at all isn't
possible, hence why all images have a width of 350px.

This commit doesn't add support for structured content in Anki yet.
Because this feature is so big, it is for the best if the work is split
across two commits.