readme: adds important message
The idea that Russians should be banned from the entirety of GitHub for
the actions of their government is absolutely contrary to the ideals of
free software and human decency. To any American, the idea that citizens
who likely had very little choice in the selection of their political
representatives and even less choice in their political decisions should
be held accountable for them is simply ridiculous. I know I would hate
it and find it unfair. I see no reason to extend those same sympathies
to Russians who now find themselves in the difficult situation of being
the targets of misdirected rage.

At the end of the day, what's most important to me is that the number of
people who care about free software continues to grow to the point that
I might one day no longer need to use anything that spies on its users
for profit. Russians and Ukrainians have both contributed a great deal
to this space, and to drive a wedge between this community because of
politics will never sit well with me.