ankiclient: added support for structured content in glossaries
This change required duplicating a lot of code from glossary label, but
it was neccessary because the differences between what Qt's rich text
engine is capabale of and QtWebEngine are capable of are fundementally
different. This version is the no-holds-barred version of what
GlossaryLabel is able to do.

Implementing this required that files be addable outside of addNote,
which means addMedia was implemented. It is just an AnkiConnect multi of
storeMediaFiles.