databasemanager: converts half-width katakana to full-width in searches
Half-width katakana is now converted to full-width before all searches.
Half-width katakana is still used in raw queries, but now it is also
converted so the maximum number of terms can be retrieved.

This commit also changes search behavior to be way more permissive for
katakana. For example, prior to this commit, a search like オ前 would
turn up nothing, but now it will turn up results for お前 as would be
expected. This sort of non-standard katakana usage is fairly common in
subtitles, so this is a nice addition.

Closes #87.