termwidget: implements recursive search on glossaries
Allows for recursive search to be done on ONLY glossaries in TermWidget.
This can be done by hoving and holding shift.

This is a massive commit that required a whole lot of work to be done to
make this feasable. A lot of the effort comes form converting singletons
into actual reusable objects. Mainly, Terms are now handled in "custom"
types, which are actually just typedef'd QSharedPointers. This is
because it makes declaring metatypes (done in main()) far easier, even
if it obfusticates what the actual types are. This should prevent any
potential UB or memory leaks that would occurr now that it's gotten so
hard to manage ownership of all these shared objects.

The work done to DefinitionWidget is more straighforward. It just keeps
a reference to it's only child (which is a child of it's parent, so it
could outlive the life of it's DefinitionParent if we're not careful),
and operates on it as if it were the only child in the tree. The child
in turn does the same to its child, and so on. This allows for infinite
recursive searched, or at least as many recursive searches as your
computer can fit in memory.

I will be adding the ability to change the search modifier, but this
commit is already way too big, so I'll leave it where it is.

Closes #80.