definitionwidget: fixes crash when clicking kanji
Due to changes in 0456eb46df7a29b6372ca75d50afbd4e1f6002a3, a crash was
introduced when hide() and show() would be called on nullptr. This is
because the stretch item in the layout of DefinitionWidget is not
actually a widget. This does a simple check that the item contains a
widget before calling hide() and show().