termwidget: fixes race condition that would prevent recursive card adds
This is a similar bug to #81, but it only affects recursive search
results. It can be triggered by adding a card and then closing the
search before the card is actually shown to be added. This is due to
DefinitionWidget getting deleted, and taking the currently working
TermWidget with it. The fix works by unparenting TermWidget when
deleteWhenReady() is called.