dictionary: greatly improves search results for conjugated words
This was something I was aware of for awhile, but I decided initially
against implementing it due to the time and memory complexity it seemed
like the algorithm would take. Even now, implementing it will almost
certainly cause a performance hit somewhere. Essentially this solution
takes everything from an iterative solution to a recursive solution
since building out MeCab queries becomes a bottom-up problem.

If I'm correct, the algorithm is a type of DFS. Maybe there is a
potential to improve performance using some sort of dynamic programming,
but it is difficult to envision since I'm dealing with a tree structure
and not an array. Maybe a map would work?

Closes #92