playeroverlay: allows osc visibility to be toggled using delete key
This isn't the way I wanted to do this, but since mpv's OSC is actually
a lua script and not a built in thing, it need to be emulated by
Memento. This means that this functionality cannot be rebound off of the
delete key which isn't exactly what I would want. As it is, this commit
works as expected.

Closes #95