playeroverlay: fixes race cond where osc could be hidden but visible
If OSC visibility was Auto, hideOverlay was called, and visibility was
quickly toggled to Visible before the overlay was hidden, the overlay
would be hidden, even though is should have been visible. This could be
fixed by simply moving the mouse, but it is still a bad thing to leave.
This fixes the issue by putting the QPropertyAnimation in a QPointer,
which allows for checking whether or not the animation is still active.