github/workflows/build: adds ci
Adds CI for all platforms. This creates installers on Windows. It
doesn't create AppBundles on macOS only because that requires certs and
it isn't possible to make them from the command line as far as I know.
It doesn't create AppImages because the build env is Ubuntu and getting
those to work with PPA's is misrable. I completely forgot how I did it
on my VM and force pushing repeatedly to try and debug the problem is
tedious.