behaviorsettings: adds osc duration timeout
Allows the user to set how long the OSC stays on screen after the mouse
is moved. The timer is interestingly in MpvWidget instead of
PlayerOverlay which may be bad form, but the reason it's in MpvWidget in
the first place is because the PlayerOverlay tracks when the cursor
disappears and then follows suit. The minimum timeout is 50ms because
anything below produces some bad UI jank.

Related to #105.