searchwidget: allows for searching from any position using modifiers
Allows a search to be executed from any position in the search string by
holding modifiers and hovering over text. Unlike in other parts of
Memento, search results are not highlighted. I chose to do this due to
the added complexity of highlighting and because I think selecting
editable text will be cause for frustration. This is because selected
text will be overwritten when typing, which probably is not desired
behavior.

I also modified Dictionary because there was a method that was no longer
useful after modifying SearchWidget.