yomidbbuilder: prevents terms from being added with same reading and exp
Checks during insertion that a term's reading and expression aren't the
same. Some dictionaries mistakenly decide to make them the same for kana
only terms which causes problems for the grouping logic. This just
changes the reading to the empty string in the case where the reading an
expression are the same.

To prevent re-adding dictionaries, the DB version was updated to 4,
which just runs a query to set all the terms that have the same reading
and expression to have their reading be the empty string. This makes it
so dictionaries don't have to be re-added. Closes #106