cmakelists: set include directory to be src
This should get the project out of relative path hell where everything
is prefixed with ../../../.. in includes. It should also make everything
more semantic since the directory structure will logically separate
includes based on what directory they are in.

The way I chose to do this probably wasn't the best, though it was the
easiest. In the root CMakeLists.txt file, I used include_directories()
and set it to use the src/ directory. The reason this is bad is because
it doesn't have target scoping, meaning it applies to all targets. CMake
has since tried to encourage target_include_directories() instead. The
issue with this is, I want the include directory be src/ for all
targets, not just some. There are a lot of targets in Memento, so things
will starts to get pretty nasty if every targets needs it's own
target_include_directories() call.