dictionarysettings: prevents ui from freezing when adding dictionaries
Makes sure not to call getDictionaries() or any other code that is
thread-safe on the UI thread. Just spawns a worker thread and emits a
signal. The signal emits a QStringList for disabled dictionaries instead
of a QSet because QSet<String> is not a registered meta type and it
isn't worth the trouble to register it.

Adds an overload to getDictionary() that adds multiple dictionaries for
the sake of only firing off dictionariesChanged() once.