playeroverlay: fixes menubar behavior when changing osc vis on windows
It was possible to get the menubar to show in fullscreen on Windows
without having "Show menubar in fullscreen" checked in settings. This
led to a non-functional menubar being presented to the user. This
changes fixes the problem by adding a method called
menuBarHandleStateChange that is called when settings or the fullscreen
state is changed. What it does it check if OSC visibility is locked to
visible, and if it is, checks if the menubar should be shown or hidden
depending on the options set by the user.