behaviorsettings: allows the default file picker location to be set
Adds the option to change the default location the Open File and Add
Subtitle File file pickers. The was previously the last directory the
user had open and that still is the default. This change allows the user
to choose from a list of preset options like Movies, Doucments, and
Home, or set their own. Closes #110