searchsettings: adds option to auto pause when hovering over subtitles
I don't want to enable this by default, since it can get annoying for
some users if they're not expecting it. It's a great option to have.

Credit to @mdatsev for the idea, not the code. I saw this when browsing
forks that actually have commits on them and thought this was a great
idea.