termwidget: allows terms to be added using kana only
Allows terms to be added using kana only. I chose to do this differently
than Yomichan partially because I always thought the design of having
two plus buttons was needlessly confusing. This makes the feature much
more explicit by using recognizable symbols in a toggle button.

Implementing this feature required adding two icons to the IconFactory,
a field to Term structs, a button in TermWidget, and some code to handle
the change in AnkiClient.

Important to to note in AnkiClient is that expression is needed for
getting the correct audio source. To allow for changing as little code
as possible, expression becomes a calculated value just like all the
other values in createAnkiNoteObject().

Closes #114.