termwidget: removes jisho link
I seriously doubt the usefulness of this feature considering I have
never used it once. The real reason I'm removing it now is because it
takes up valuable horizontal space in TermWidget. What I'm afraid of is
a really long term being cut off because it can't fit in the space left
for it.

If anyone has a problem with this change let me know by commenting on
this commit or making an issue. The master branch is a development
branch, and I'm always open to feedback regarding changes.