behaviorsettings: allow letting mpv control cursor visibility
Add the option to have Memento let mpv control the visibility of the
cursor according to --cursor-autohide and --cursor-autohide-fs-only. The
reason this is not set as default is because the cursor already has the
desireable behavior of autohiding together with the OSC. This usecase
will be fairly niche in comparison.

To implement this, two new wrapper classes around QTimer had to be
created, OSCTimer and MpvTimer.

OSCTimer simply implements the current behavior of hiding the cursor
with the OSC. The implementation is fairly simple.

MpvTimer implements a state machine that adhears to the logic of
--cursor-autohide and --cursor-autohide-fs-only. Due to this, the logic
is long and complicated since there are a ton of different states to
consider. This is my justification for seperating the implementation
into its own class.

Closes #116