searchsettings: add vertical bar glossary style
The previous name for this option asked whether or not to use bullet
points as a binary choice. This new version now has a combobox that
allows for bullet points, line breaks, or vertical bars.

Doing this required changing the settings version to 3 and then updating
the original setting to deal with the new GlossaryStyle enum.

Related to #117