util/subtitleparser: fix subtitle parsing issue depending on uri scheme
cd779a14c7634c5e74e7969a443cc0a90b2a90ed stopped wrapping the path in a
QUrl and instead started passing the raw path. This seemed to fix the
issue with the parser in the case I was concerned about, but ended up
breaking the parser elsewhere.

Essentially when a subtitle file is dropped on the player, the scheme
that is passed to mpv is a file:// URI. mpv will then pass this back to
Memento as it was given to it. This included the file:// URI. When this
is passed to QFile, it can't cope with this and assumes the file doesn't
exist.

Likewise, when mpv loads a local file on its own (not having it
dropped into Memento), there is no file:// URI returned to Memento by
mpv. When wrapped in a QUrl object, this will fail to parse and thus
when passing it to QFile it will fail to open the file.

To fix this, both situations are now taken into account. The path is
wrapped in a QUrl, checked if it is a local file, and then a QFile is
created from the raw path or the toLocalFile of the QUrl depending on
which is necessary.