gui/widgets/overlay/playermenu: fix menus staying open on osc hide
It was possible for QMenus to stay open even when the PlayerMenu/Overlay
would be hidden. This fixes the issue by looping over every QMenu and
closing each one so that none are open.

The bug can be triggered by opening a menu, and then settings the OSC
visibility to hidden using the delete key. The menu will stay open even
though the OSC is hidden.