gui/widgets/definition/termwidget: fix open in anki not finding kana
If adding a term using kana instead of kanji, the open in Anki button
would fail to function as it only searches for the expression of the
term, not the reading. This fixes the issue by using a query that looks
for both the expression and reading when clicking open in Anki.

Closes #126