anki/ankiclient: fix notesaddable returning wrong value for kana notes
If a note was previously added as a reading when it had an expression,
notesaddble would mark it as addable even though it shouldn't be. This
fixes the issue in a naive, inefficient way, but it requires little code
change, and I'm not sure there's actually a better way.

The way this was fixed involved creating two notes per term, one with
readingAsExpression set to false for the term, and one with it set to
true. Both notes would then be sent to AnkiConnect to evaluated. The
result of this would be intercepted by an AnkiReply privately created in
notesAddable that would aggregate the two different addable values into
one addable value, and then pass that on to the caller's AnkiReply.

Closes #127