memento: add ocr support with manga-ocr
Add OCR support using manga-ocr and libmocr. This is a basic start for
this feature. Later I would like to add user selectable models, and the
ability to disable OCR entirely. Importantly, this solution gets past
the license incompatibility between GPLv2 and Apache 2.0 since libmocr
is the only software Memento links to. Because libmocr is under LGPLv3
(or really any license I like, I do own the copyright), Memento is not
obligated to follow the terms of manga-ocr's license so long as any
binaries I redistribute are dynamically linked to libmocr.

The downside of this feature is that it is very, very, very heavy.
PyTorch is nearly a gigabyte, Python is probably several hundred
megabytes, and the model it uses is very large as well. All together
this bloats the original install size of Memento by several times. In
addition, it is hard to ship additional Python libraries along side the
binaries. I'm certain it is possible, but I'm not yet sure how to do it.
For all these reason, it is almost certain that no binaries will be
provided with OCR support.

As for the technical aspects of this PR, libmocr is pulled in at
configure time using FetchContent and is dynamically linked to Memento.
If users choose to redistribute binaries of Memento, that will be
completely legal.

PlayerOverlay was minorly changed to now how two modes, the OCROverlay
and the standard overlay. The standard overlay is contained in a widget
and it visible normally. The OCROverlay will replace it when OCR mode is
entered. This mode can be escaped by pressing ESC or selecting an area
to run OCR on.

The only way to toggle OCR mode is by clicking the eye icon in the
player controls. I will likely add a key shortcut later, but for now
this is the only way to enter OCR mode.

When a search is triggered, the SearchWidget is shown if it isn't
already visible and the text is immediately searched. This is the most
elegant solution to integrating OCR and works quite well.

Lastly, the GitHub workflows have been updated to build with OCR support
on. Since manga-ocr isn't needed at build time, none of the workflows
install it. To save on time, non-OCR versions aren't built.