anki/ankiclient: fix notesaddable failing for kana only terms
This bug would occur because notesAddable unconditionally generated
notes for terms without a reading, aka kana only terms. This means it
would check if a term with an empty {expression} marker was addable,
which it of course would not be if {expression} is used as the note key.
This fixes the issue by not generating notes for terms without a
reading. When the results are returned by Anki, notesAddable are
aggregated into one value by checking if the current term for the
current value has a reading or not.

Closes #133