cmakelists: set macosx_deployment_target to 11 for release builds
If the RELEASE_BUILD option is set on macOS, the macosx_deployment_taget
will be set to Big Sur. This will hopefully allow me to keep my personal
machine up to date while providing builds compatible with older OSes.
This option is only set when RELEASE_BUILD=ON since it is possible to
run Memento on older versions of macOS, so I don't want to prevent
normal users from compiling it themselves on something like 10.15.
RELASE_BUILD is an undocumented option, so users will only set it if
they know what they are doing.