version: change version to v1.0.0
Version 1.0.0 doesn't mean anything special. I just consider this
release to be feature complete and stable as far as Memento goes. Thus
the -beta suffix is now dropped from versioning. There are of course
still things that can be added, but in its current state Memento is good
enough to be considered complete software.